Add files via upload
Legal documents related to ABPMJAY were processed using Apache Tika to extract information from web pages. In the Python work, a rule-based approach was designed to extract information in triple form(subject,verb,object) from both "simple" as well as "complex" sentences, leveraging the positions of entities in the sentence. Python lists played a crucial role in this extraction process.

For more advanced linguistic analysis, the "encore_web_md" pretrained model was utilized for Parts of Speech (POS) tagging and Dependency Parsing of sentences through deep learning techniques.

To organize and represent the extracted information, Neo4j was employed to construct a summary 
knowledge graph based on the generated triplets. The Cypher Query Language (CQL) was then utilized for the modification, deletion, and insertion of nodes in the knowledge graph.

#Python #Apache Tika #Deep Learning #Neo4j #Knowledge Graphs.